IAgent10_n8n_csv/requirements.txt
fastapi
uvicorn
openai
pyzipper
tqdm
tiktoken
langchain
openai>=1.14.3
faiss-cpu
pandas
python-dotenv
streamlit
requests
supabase
sentence-transformers



